requirements.txt
Pillow
pydub
numpy
